+++ sport-log-types/migrations/2021-07-13-151521_sharing/up.sql
@@ -11,8 +11,8 @@ create trigger set_timestamp before update on "group"
 
 create table group_user (
     id bigint primary key,
-    group_id integer not null references "group" on delete cascade,
-    user_id integer not null references "user" on delete cascade,
+    group_id bigint not null references "group" on delete cascade,
+    user_id bigint not null references "user" on delete cascade,
     last_change timestamptz not null default now(),
     deleted boolean not null default false,
     unique (group_id, user_id, deleted)
@@ -23,8 +23,8 @@ create trigger set_timestamp before update on group_user
 
 create table shared_metcon_session (
     id bigint primary key,
-    group_id integer not null references "group" on delete cascade,
-    metcon_session_id integer not null references metcon_session on delete cascade,
+    group_id bigint not null references "group" on delete cascade,
+    metcon_session_id bigint not null references metcon_session on delete cascade,
     last_change timestamptz not null default now(),
     deleted boolean not null default false,
     unique (group_id, metcon_session_id, deleted)
@@ -35,8 +35,8 @@ create trigger set_timestamp before update on shared_metcon_session
 
 create table shared_strength_session (
     id bigint primary key,
-    group_id integer not null references "group" on delete cascade,
-    strength_session_id integer not null references strength_session on delete cascade,
+    group_id bigint not null references "group" on delete cascade,
+    strength_session_id bigint not null references strength_session on delete cascade,
     last_change timestamptz not null default now(),
     deleted boolean not null default false,
     unique (group_id, strength_session_id, deleted)
@@ -47,8 +47,8 @@ create trigger set_timestamp before update on shared_strength_session
 
 create table shared_cardio_session (
     id bigint primary key,
-    group_id integer not null references "group" on delete cascade,
-    cardio_session_id integer not null references cardio_session on delete cascade,
+    group_id bigint not null references "group" on delete cascade,
+    cardio_session_id bigint not null references cardio_session on delete cascade,
     last_change timestamptz not null default now(),
     deleted boolean not null default false,
     unique (group_id, cardio_session_id, deleted)
@@ -59,8 +59,8 @@ create trigger set_timestamp before update on shared_cardio_session
 
 create table shared_diary (
     id bigint primary key,
-    group_id integer not null references "group" on delete cascade,
-    diary_id integer not null references diary on delete cascade,
+    group_id bigint not null references "group" on delete cascade,
+    diary_id bigint not null references diary on delete cascade,
     last_change timestamptz not null default now(),
     deleted boolean not null default false,
     unique (group_id, diary_id, deleted)
